add PPOTrainer base class
