+++ verl/trainer/config/_generated_ppo_megatron_trainer.yaml
@@ -899,6 +899,15 @@ trainer:
   max_critic_ckpt_to_keep: null
   ray_wait_register_center_timeout: 300
   device: cuda
+  use_v1: false
+  v1:
+    trainer_mode: sync
+    sync: {}
+    colocate_async:
+      num_warmup_batches: 1
+    separate_async:
+      num_warmup_batches: 4
+      parameter_sync_step: 4
 global_profiler:
   _target_: verl.utils.profiler.ProfilerConfig
   tool: null
