+++ verl/trainer/ppo/v1/trainer_colocate_async.py
@@ -1,0 +1,52 @@
+# Copyright 2024 Bytedance Ltd. and/or its affiliates
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+import logging
+import os
+
+from verl.trainer.ppo.v1.trainer_base import PPOTrainer, register_trainer
+from verl.utils.debug import marked_timer
+
+logger = logging.getLogger(__name__)
+logger.setLevel(os.getenv("VERL_LOGGING_LEVEL", "INFO"))
+
+
+@register_trainer("colocate_async")
+class PPOTrainerColocateAsync(PPOTrainer):
+    """Asynchronous PPO trainer
+    1. Trainer and rollout are colocated.
+    2. Partial rollout is enabled.
+    """
+
+    def on_init_end(self):
+        # update weights after loading checkpoint
+        self.checkpoint_manager.update_weights(self.global_steps)
+
+    def on_train_begin(self):
+        num_warmup_batches = self.config.trainer.v1.colocate_async.num_warmup_batches
+        for _ in range(num_warmup_batches):
+            self.add_batch_to_generate()
+        logger.info(f"Added {num_warmup_batches} warmup batches to the agent loop manager")
+
+    def on_step_end(self):
+        with marked_timer("update_weights", self.timing_raw, color="red"):
+            # wake up all replicas to update weights
+            self.checkpoint_manager.update_weights(self.global_steps)
+            # resume generation
+            self.checkpoint_manager.resume_generation_replicas()
+
+    def on_sample_end(self):
+        # abort all unfinished requests and pause generation
+        self.checkpoint_manager.abort_replicas()
+        # sleep all replicas to discard weights and kv cache
+        self.checkpoint_manager.sleep_replicas()
