+++ verl/trainer/ppo/v1/utils.py
@@ -1,0 +1,92 @@
+# Copyright 2024 Bytedance Ltd. and/or its affiliates
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+from typing import Any
+
+import torch
+
+from verl.protocol import DataProto
+from verl.trainer.ppo import core_algos
+from verl.trainer.ppo.ray_trainer import compute_advantage
+
+
+def compute_advantage_for_multi_trajectories(
+    data: DataProto,
+    batch_keys: list[str],
+    adv_estimator,
+    gamma: float = 1.0,
+    lam: float = 1.0,
+    num_repeat: int = 1,
+    norm_adv_by_std_in_grpo: bool = True,
+    config: Any = None,
+) -> DataProto:
+    """Compute GRPO advantages from each session's final output. For non-GRPO
+    estimators, such as GAE, are delegated to the original compute_advantage() unchanged.
+
+    For GRPO, only the final output in each ``{uid}_{session_id}`` group participates
+    in advantage computation, and the result is broadcast to the other outputs in
+    the same session. Sessions whose AgentLoop returns ``None`` simply do not appear
+    in ``batch_keys``. Non-GRPO estimators, such as GAE, are delegated to the
+    original ``compute_advantage()`` unchanged.
+    """
+    if adv_estimator != core_algos.AdvantageEstimator.GRPO:
+        return compute_advantage(
+            data,
+            adv_estimator=adv_estimator,
+            gamma=gamma,
+            lam=lam,
+            num_repeat=num_repeat,
+            norm_adv_by_std_in_grpo=norm_adv_by_std_in_grpo,
+            config=config,
+        )
+
+    # final session of each agent loop: {uid}_{session_id} => (index, row_index)
+    final_sessions: dict[str, tuple[int, int]] = {}
+    row_session_keys = []
+    for i, key in enumerate(batch_keys):
+        fields = key.rsplit("_", 2)
+        assert len(fields) == 3, f"Unexpected key format: {key}"
+        uid, session_id, index = fields[0], fields[1], int(fields[2])
+        session_key = f"{uid}_{session_id}"
+        row_session_keys.append(session_key)
+        if session_key not in final_sessions or final_sessions[session_key][0] < index:
+            final_sessions[session_key] = (index, i)
+
+    # final session indices in batch data
+    final_indices = []
+    session_key_to_local_index = {}
+    for session_key, (_, row_index) in final_sessions.items():
+        final_indices.append(row_index)
+        session_key_to_local_index[session_key] = len(final_indices) - 1
+    row_to_local_index = [session_key_to_local_index[session_key] for session_key in row_session_keys]
+
+    # select final sessions from batch data for group relative advantage computation
+    final_data = compute_advantage(
+        data.select_idxs(final_indices),
+        adv_estimator=adv_estimator,
+        gamma=gamma,
+        lam=lam,
+        num_repeat=num_repeat,
+        norm_adv_by_std_in_grpo=norm_adv_by_std_in_grpo,
+        config=config,
+    )
+    first_nnz_indices = final_data.batch["response_mask"].argmax(dim=1)
+    final_scores = final_data.batch["advantages"][torch.arange(len(final_data)), first_nnz_indices]
+
+    # scatter final scores to all rows in batch data
+    scores = final_scores[row_to_local_index]
+    scores = scores.unsqueeze(-1) * data.batch["response_mask"]
+
+    data.batch["advantages"] = scores
+    data.batch["returns"] = scores
+    return data
